fix: soname must be set after rpath
dpkg-shlibdeps is sensitive to the order of the SONAME and RPATH entries
in the shared library files. It expects the SONAME after the RPATH.
patchelf would put the entries in the opposite order if the SONAME and
the RPATH are set in the same command, hence we need to run patchelf
twice to define the order as expected by dpkg-shlibdeps.